# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: RISE API
  version: 1.0.0
  description: RISE — Autonomous Incident Remediation System API Specification
servers:
  - url: https://api.rise.internal/api/v1
    description: Internal API Gateway
security:
  - BearerAuth: []
paths:
...
  /webhooks/cloudwatch:
    post:
      summary: CloudWatch SNS webhook receiver
      operationId: cloudwatchWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponseEnvelope'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /webhooks/alertmanager:
    post:
      summary: Alertmanager webhook receiver
      operationId: alertmanagerWebhook
      security: []
      parameters:
        - name: X-RISE-Secret
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponseEnvelope'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /webhooks/github:
    post:
      summary: GitHub event webhook receiver
      operationId: githubWebhook
      security: []
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponseEnvelope'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /webhooks/slack:
    post:
      summary: Slack interactive payload webhook receiver
      operationId: slackWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponseEnvelope'
        '422':
          $ref: '#/components/responses/422ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Meta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - RISK_POLICY_VIOLATION
            - APPROVAL_EXPIRED
            - ACTION_PLAN_CHANGED
            - RESOURCE_LOCKED
            - INTEGRATION_UNAVAILABLE
            - RATE_LIMITED
            - UNAUTHORIZED
            - NOT_FOUND
        message:
          type: string
        details:
          type: object
    ErrorEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: object
          nullable: true
          default: null
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    WebhookResponseData:
      type: object
      required: [received]
      properties:
        received:
          type: boolean
    WebhookResponseEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/WebhookResponseData'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
...
  responses:
...
    422ValidationError:
      description: Unprocessable Entity - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "VALIDATION_ERROR"
              message: "Request body failed schema validation"
              details: {}
...